[2.3.1] RKE Template Bugs (#3446)
* check if k8s version is user override or patchVersion override

rancher/rancher#23413
rancher/rancher#23415

In the event that the cluster template creator selected a k8s version that was a
patch release version when we came in to clone the revision the UI would show
the question as an override in which case the `satisfies` param was dropped. Now
when a clone happens we check to see if its the patch and has satisifies and in
this case we "remove" the override because its not technically one. The logic on
the cluster template save will ensure the new "override" is added on save with
the correct version.

* ensure clusterTemplateQuestions is never null, always an empty array

rancher/rancher#23404

I could never repro the issue above but when looking at the code and the image
in the issue it looked as though the clusterTemplateQuestions variable was null
when trying to each over it so I've added a new check that ensures the questions
are not null so the UI doesn't blow up.